Create new GuestLayout and refactor guest pages to utilize
