+++ src/pages/account/login.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import gql from 'graphql-tag';
 import { Mutation } from 'react-apollo'
-import { Link, navigate, replace } from 'gatsby'
+import { Link, navigate } from 'gatsby'
 import ContextConsumer from '../../layouts/context'
-import AuthenticationWrapper from '../../components/account/AuthenticationWrapper'
+import GuestLayout from '../../components/account/GuestLayout'
 
 const CUSTOMER_LOGIN = gql`
 mutation customerAccessTokenCreate($input: CustomerAccessTokenCreateInput!) {
@@ -117,13 +117,9 @@ class Login extends React.Component {
         )
 
         return (
-            <AuthenticationWrapper>
-                {({ isAuthenticated }) => (
-                    (isAuthenticated)
-                        ? replace(`/account`)
-                        : pageContent
-                )}
-            </AuthenticationWrapper>
+            <GuestLayout>
+                {pageContent}
+            </GuestLayout>
         )
     }
 }
